Update command to be `spice refresh dataset-name`
